# resource/Contatti_db/Rubrica_telefonica.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE Rubrica_telefonica.js PLEASE EDIT ../custom/Rubrica_telefonicaCustom.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
const app = require('../../app.js');
const db_Contatti_db = require('../../db/Contatti_db_schema.js')
const properties = require('../../properties.js');
const handleError = require('../../security/util.js').handleError;
require('./custom/Rubrica_telefonicaCustom.js');

/*
 * SCHEMA DB Rubrica_telefonica
 * 
	{
		cognome: {
			type: 'String'
		},
		nome: {
			type: 'String'
		},
		//RELATIONS
		
		
		//EXTERNAL RELATIONS
		
		
	}
 * 
 */



//CRUD METHODS


/**
 * Rubrica_telefonicaService.create
 *   @description CRUD ACTION create
 *
 */
app.post(properties.api + '/rubrica_telefonicas', function (req, res) {
	obj = new db_Contatti_db.Rubrica_telefonica(req.body);
	obj.save(function (err) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * Rubrica_telefonicaService.delete
 *   @description CRUD ACTION delete
 *   @param ObjectId id Id
 *
 */
app['delete'](properties.api + '/rubrica_telefonicas/:id', function (req, res) {
	db_Contatti_db.Rubrica_telefonica.findByIdAndRemove(req.params.id, function (err) {
		  if (err) return handleError(err, res);
		  res.send(err);
	});
});

/**
 * Rubrica_telefonicaService.get
 *   @description CRUD ACTION get
 *   @param ObjectId id Id 
 *
 */
app.get(properties.api + '/rubrica_telefonicas/:id', function (req, res) {
	db_Contatti_db.Rubrica_telefonica.findOne({_id:req.params.id}).exec(function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * Rubrica_telefonicaService.list
 *   @description CRUD ACTION list
 *
 */
app.get(properties.api + '/rubrica_telefonicas', function (req, res) {
	db_Contatti_db.Rubrica_telefonica.find().exec(function (err, list) {
		if (err) return handleError(err, res);
		res.send(list);
	});
});

/**
 * Rubrica_telefonicaService.update
 *   @description CRUD ACTION update
 *   @param ObjectId id Id
 *
 */
app.post(properties.api + '/rubrica_telefonicas/:id', function (req, res) {
	db_Contatti_db.Rubrica_telefonica.findByIdAndUpdate(req.params.id, req.body, {'new': true}, function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});


/*
 * CUSTOM SERVICES
 * 
 *	These services will be overwritten and implemented in  Custom.js
 */